Fix travis
The matrix property was not working properly. For now, the matrix has
been refactored to 2 scripts of the same stage. This is not ideal as it
will lead to code duplication in the future. One possible workaround to
duplication is to have variables/references of the configuration we want
and reference it when necessary. One possible _more definitive_ solution
would be to understand why matrix is not working and fix it. (I tried to
address this, but due to time constraints I had to postpone this task)